+++ core/delegation/config.go
@@ -13,7 +13,7 @@
 package delegation
 
 import (
-	"code.vegaprotocol.io/vega/core/config/encoding"
+	"code.vegaprotocol.io/vega/libs/config/encoding"
 	"code.vegaprotocol.io/vega/logging"
 )