+++ libs/pprof/pprof.go
@@ -15,6 +15,7 @@ package pprof
 import (
 	"fmt"
 	"net/http"
+
 	// import pprof globally because it's used to init the package
 	// and this comment is mostly here as well in order to make
 	// golint very many much happy.
@@ -25,7 +26,7 @@ import (
 	"runtime/pprof"
 	"time"
 
-	"code.vegaprotocol.io/vega/core/config/encoding"
+	"code.vegaprotocol.io/vega/libs/config/encoding"
 	vgfs "code.vegaprotocol.io/vega/libs/fs"
 	"code.vegaprotocol.io/vega/logging"
 )